Move getProjectNameFromPath function to remove circular dependency
