+++ src/pfe/file-watcher/server/src/projects/projectUtil.ts
@@ -104,7 +104,7 @@ export async function containerCreate(operation: Operation, script: string, comm
     const event = "projectCreation";
     const projectLocation = operation.projectInfo.location;
     const projectID = operation.projectInfo.projectID;
-    const projectName = getProjectNameFromPath(projectLocation);
+    const projectName = utils.getProjectNameFromPath(projectLocation);
     const projectType = operation.projectInfo.projectType;
     if (projectList.indexOf(projectID) === -1)
         projectList.push(projectID);
@@ -211,7 +211,7 @@ export async function containerUpdate(operation: Operation, script: string, comm
 
     const projectLocation = operation.projectInfo.location;
     const projectID = operation.projectInfo.projectID;
-    const projectName = getProjectNameFromPath(projectLocation);
+    const projectName = utils.getProjectNameFromPath(projectLocation);
     const projectType = operation.projectInfo.projectType;
     logger.logProjectInfo("Updating container for " + operation.projectInfo.projectType + " project " + projectLocation, projectID, projectName);
     operation.containerName = await getContainerName(operation.projectInfo);
@@ -316,7 +316,7 @@ export async function containerUpdate(operation: Operation, script: string, comm
 async function executeBuildScript(operation: Operation, script: string, args: Array<string>, event: string): Promise<void> {
     const projectID = operation.projectInfo.projectID;
     const projectLocation = operation.projectInfo.location;
-    const projectName = getProjectNameFromPath(projectLocation);
+    const projectName = utils.getProjectNameFromPath(projectLocation);
     const projectInfo = {
         operationId: operation.operationId,
         projectID: operation.projectInfo.projectID
@@ -612,7 +612,7 @@ export async function getProjectMavenSettings(projectInfo: ProjectInfo): Promise
 export async function getProjectLogs(projectInfo: ProjectInfo): Promise<ProjectLog> {
     const projectID = projectInfo.projectID;
     const projectLocation = projectInfo.location;
-    const projectName = getProjectNameFromPath(projectLocation);
+    const projectName = utils.getProjectNameFromPath(projectLocation);
     const projectType = projectInfo.projectType;
     const projectLogDir = await logHelper.getLogDir(projectID, projectName);
     const logDirectory = path.join(projectConstants.projectsLogDir, projectLogDir);
@@ -655,7 +655,7 @@ export async function getProjectLogs(projectInfo: ProjectInfo): Promise<ProjectL
 export async function containerDelete(projectInfo: ProjectInfo, script: string): Promise<void> {
 
     const projectID = projectInfo.projectID;
-    const projectName = getProjectNameFromPath(projectInfo.location);
+    const projectName = utils.getProjectNameFromPath(projectInfo.location);
     const containerName = await getContainerName(projectInfo);
     const imagePushRegistry = projectInfo.deploymentRegistry;
     logger.logProjectInfo("containerDelete: Kill running processes and remove container... ", projectID, projectName);
@@ -725,7 +725,7 @@ export function getLogName(projectID: string, projectLocation: string): string {
     const hash = crypto.createHash("sha1", <TransformOptions>"utf8").update(projectLocation);
 
     let logName = projectConstants.containerPrefix + projectID + "-" + hash.digest("hex");
-    const projectName = getProjectNameFromPath(projectLocation);
+    const projectName = utils.getProjectNameFromPath(projectLocation);
 
     if (process.env.IN_K8 === "true" && logName.length > 53) {
         logName = logName.substring(0, 53);
@@ -751,7 +751,7 @@ export function getDefaultContainerName(projectID: string, projectLocation: stri
     }
 
     // Sanitize project name to ensure project name only supports lower case letter and number
-    const projectNameOrigin: string = getProjectNameFromPath(projectLocation);
+    const projectNameOrigin: string = utils.getProjectNameFromPath(projectLocation);
     const letterNumber: RegExp = /[A-Za-z0-9]/;
     const upperCaseLetter: RegExp = /[A-Z]/;
     const defaultProjectName: string = "cw";
@@ -1249,7 +1249,7 @@ export async function runScript(projectInfo: ProjectInfo, script: string, comman
     const containerName = await getContainerName(projectInfo);
     const logName = getLogName(projectInfo.projectID, projectInfo.location);
     const logDir = await logHelper.getLogDir(projectInfo.projectID, projectInfo.projectName);
-    const projectName = getProjectNameFromPath(projectInfo.location);
+    const projectName = utils.getProjectNameFromPath(projectInfo.location);
     let args = [projectInfo.location, LOCAL_WORKSPACE, projectID, command, containerName, String(projectInfo.autoBuildEnabled), logName, projectInfo.startMode,
         projectInfo.debugPort, "NONE", logDir];
 
@@ -1290,7 +1290,7 @@ export async function buildAndRun(operation: Operation, command: string): Promis
 
     const projectLocation = operation.projectInfo.location;
     const projectID = operation.projectInfo.projectID;
-    const projectName = getProjectNameFromPath(projectLocation);
+    const projectName = utils.getProjectNameFromPath(projectLocation);
 
     if (projectList.indexOf(projectID) === -1)
         projectList.push(projectID);
@@ -1446,7 +1446,7 @@ export async function buildAndRun(operation: Operation, command: string): Promis
  */
 async function containerBuildAndRun(event: string, buildInfo: BuildRequest, operation: Operation): Promise<void> {
     const normalizedProjectLocation = path.resolve(buildInfo.projectLocation);
-    const projectName = getProjectNameFromPath(normalizedProjectLocation);
+    const projectName = utils.getProjectNameFromPath(normalizedProjectLocation);
     const logDir = await logHelper.getLogDir(buildInfo.projectID, projectName);
     const dockerBuildLog = path.resolve(buildInfo.projectLocation + "/../.logs/" + logDir, logHelper.buildLogs.dockerBuild + logHelper.logExtension);
     if (process.env.IN_K8 === "true") {
@@ -1668,7 +1668,7 @@ async function containerBuildAndRun(event: string, buildInfo: BuildRequest, oper
  */
 async function runLocalContainer(buildInfo: BuildRequest): Promise<void> {
     const normalizedProjectLocation = path.resolve(buildInfo.projectLocation);
-    const projectName = getProjectNameFromPath(normalizedProjectLocation);
+    const projectName = utils.getProjectNameFromPath(normalizedProjectLocation);
     const logDir = await logHelper.getLogDir(buildInfo.projectID, projectName);
     const appLog = path.resolve(buildInfo.projectLocation + "/../.logs/" + logDir, logHelper.appLogs.app + logHelper.logExtension);
     try {
@@ -1799,7 +1799,7 @@ export async function isApplicationPodUp(buildInfo: BuildRequest, projectName: s
 export async function removeProject(projectInfo: ProjectInfo): Promise<void> {
 
     const projectID = projectInfo.projectID;
-    const projectName = getProjectNameFromPath(projectInfo.location);
+    const projectName = utils.getProjectNameFromPath(projectInfo.location);
     const containerName = await getContainerName(projectInfo);
     logger.logProjectInfo("removeProject: Kill running processes and remove container... ", projectID, projectName);
     logger.logProjectInfo("Project ID:        " + projectInfo.projectID, projectID, projectName);
@@ -1903,7 +1903,7 @@ async function getPODInfoAndSendToPortal(operation: Operation, event: string = "
     const projectInfo = operation.projectInfo;
     const projectLocation = projectInfo.location;
     const projectID = projectInfo.projectID;
-    const projectName = getProjectNameFromPath(projectLocation);
+    const projectName = utils.getProjectNameFromPath(projectLocation);
     const keyValuePair: UpdateProjectInfoPair = {
         key: "buildRequest",
         value: false
@@ -2203,26 +2203,8 @@ export async function updateDetailedAppStatus(projectID: string, ip: string, por
 export async function exposeOverIngress(projectInfo: ProjectInfo, appPort?: number): Promise<void> {
     if (process.env.IN_K8) {
         const projectID = projectInfo.projectID;
-        const projectName = getProjectNameFromPath(projectInfo.location);
+        const projectName = utils.getProjectNameFromPath(projectInfo.location);
         projectInfo.appBaseURL = await kubeutil.exposeOverIngress(projectID, projectName, projectInfo.isHttps, appPort, projectInfo.appBaseURL);
         await projectsController.saveProjectInfo(projectID, projectInfo, true);
     }
-}
-
-/**
- * Get the projectName from a path
- *
- * @param location The project's location
- */
-export function getProjectNameFromPath(location: string): string {
-    const splitPaths = location.split("/");
-    const cwIndex = splitPaths.indexOf("codewind-workspace");
-    if (cwIndex === -1) {
-        logger.logError("Unable to get project name from path: codewind-workspace isn't in the path");
-        // Fall back to old method if codewind-workspace isn't in the path
-        return path.basename(location);
-    }
-    // Project name is the directory after codewind-workspace
-    const projectName = splitPaths[cwIndex + 1];
-    return projectName;
 }
